Add --dryrun option, colors, and tabular output
